Wallet resume Improvements
