+++ src/NEO.Sync/NEO.Sync.csproj
@@ -22,4 +22,8 @@
     <PackageReference Include="NeoModules.RPC" Version="1.2.1" />
     <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="3.1.4" />
   </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Migrations\" />
+  </ItemGroup>
 </Project>
