+++ src/NEO.Sync/Program.cs
@@ -28,7 +28,7 @@ namespace NEO.Api.Worker
     public class Startup
     {
         public void ConfigureServices(IServiceCollection services)
-            => services.AddDbContext<EvoLuaContext>();
+            => services.AddDbContext<NeoContext>();
     }
 
 }
